Update event interface to use io.Reader over simple strings
Requiring a string for the event payload necessitates keeping the entire
payload event in memory. It would be better from a memory-usage POV if
we could stream the data via an io.Reader.

While the interface in this package is changing, the reading and parsing
behavior isn't. In other words, any event generated by this library (at
least for now), is still reading the entire event payload into memory
when it comes across the wire.

However, this change enables the LaunchDarkly SDK to create an event
from within our relay proxy which takes advantage of streaming, thereby
reducing the memory footprint of that application quite a lot.